move prefix to testutils
