+++ testutils/backend.go
@@ -155,7 +155,7 @@ func (b *MemoryBackend) ListObjects(ctx context.Context, accessKeyID *string, bu
 	var lastMatchedPart string
 
 	for _, obj := range objects {
-		match := prefix.Match(obj.name)
+		match := match(prefix, obj.name)
 		switch {
 		case match == nil:
 			continue
